+++ scanpipe/tests/test_api.py
@@ -628,7 +628,7 @@ class ScanPipeAPITest(TransactionTestCase):
 
     def test_scanpipe_api_serializer_get_serializer_fields(self):
         self.assertEqual(31, len(get_serializer_fields(DiscoveredPackage)))
-        self.assertEqual(27, len(get_serializer_fields(CodebaseResource)))
+        self.assertEqual(28, len(get_serializer_fields(CodebaseResource)))
 
         with self.assertRaises(LookupError):
             get_serializer_fields(None)