+++ scanpipe/tests/test_pipes.py
@@ -1222,3 +1222,17 @@ class ScanPipePipesTransactionTest(TransactionTestCase):
         make_codebase_resource(p1, resource_location)
         self.assertEqual(1, p1.codebaseresources.count())
         self.assertEqual(0, p1.projecterrors.count())
+
+    def test_scanpipe_add_to_package(self):
+        project1 = Project.objects.create(name="Analysis")
+        codebase_resource = CodebaseResource.objects.create(
+            project=project1,
+            path="filename.ext",
+        )
+        package1 = update_or_create_package(project1, package_data1)
+        self.assertFalse(codebase_resource.for_packages)
+        scancode.add_to_package(package1.package_uid, codebase_resource, project1)
+        for_packages = codebase_resource.for_packages
+        self.assertEqual(len(for_packages), 1)
+        result_purl = for_packages[0]
+        self.assertEqual(result_purl, package1.purl)